requirements.txt
openai==0.27.0
python-dotenv==0.19.0